### 2022-06-04 a0368c4
Updated leverage explanation
