+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = TLViz
+name = tensorly-viz
 version = 0.1.1
 license = "MIT license",
 description = Package to visualise component-based decomposition models such as PCA and PARAFAC
