### 2022-06-04 ff8a93a
Updated installation instructions
