+++ docs/installation.rst
@@ -12,7 +12,7 @@ To install TLViz, run this command in your terminal:
 
 .. code-block:: console
 
-    pip install tlviz
+    pip install tensorly-viz
 
 Alternatively, to get the latest development version of TLViz, run:
 
