+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = tensorly-viz
-version = 0.1.2
+version = 0.1.3
 license = "MIT license",
 description = Package to visualise component-based decomposition models such as PCA and PARAFAC
 long_description = file: README.rst
