+++ stable/_sources/auto_examples/plot_selecting_aminoacids_components.rst.txt
@@ -480,7 +480,7 @@ we find that one of the components have negative, and therefore unphysical, comp
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  11.397 seconds)
+   **Total running time of the script:** ( 0 minutes  13.198 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_selecting_aminoacids_components.py:
