+++ stable/_sources/auto_examples/plot_working_with_xarray.rst.txt
@@ -497,7 +497,7 @@ Storing this dataset as an xarray data array
     Coordinates:
       * month        (month) &lt;U3 &#x27;Jan&#x27; &#x27;Feb&#x27; &#x27;Mar&#x27; &#x27;Apr&#x27; ... &#x27;Sep&#x27; &#x27;Oct&#x27; &#x27;Nov&#x27; &#x27;Dec&#x27;
       * day-of-week  (day-of-week) &lt;U3 &#x27;Mon&#x27; &#x27;Tue&#x27; &#x27;Wed&#x27; &#x27;Thu&#x27; &#x27;Fri&#x27; &#x27;Sat&#x27; &#x27;Sun&#x27;
-      * hour         (hour) int64 1 2 3 4 5 6 7 8 9 ... 16 17 18 19 20 21 22 23 24</pre><div class='xr-wrap' style='display:none'><div class='xr-header'><div class='xr-obj-type'>xarray.DataArray</div><div class='xr-array-name'></div><ul class='xr-dim-list'><li><span class='xr-has-index'>hour</span>: 24</li><li><span class='xr-has-index'>day-of-week</span>: 7</li><li><span class='xr-has-index'>month</span>: 12</li></ul></div><ul class='xr-sections'><li class='xr-section-item'><div class='xr-array-wrap'><input id='section-54bfe937-b6fb-46a6-930a-02ebf4ecc921' class='xr-array-in' type='checkbox' checked><label for='section-54bfe937-b6fb-46a6-930a-02ebf4ecc921' title='Show/hide data repr'><svg class='icon xr-icon-database'><use xlink:href='#icon-database'></use></svg></label><div class='xr-array-preview xr-preview'><span>11 2 11 13 14 12 14 8 12 11 12 12 6 ... 9 13 13 5 14 12 14 9 12 9 17 8</span></div><div class='xr-array-data'><pre>array([[[11,  2, 11, ..., 11, 12, 12],
+      * hour         (hour) int64 1 2 3 4 5 6 7 8 9 ... 16 17 18 19 20 21 22 23 24</pre><div class='xr-wrap' style='display:none'><div class='xr-header'><div class='xr-obj-type'>xarray.DataArray</div><div class='xr-array-name'></div><ul class='xr-dim-list'><li><span class='xr-has-index'>hour</span>: 24</li><li><span class='xr-has-index'>day-of-week</span>: 7</li><li><span class='xr-has-index'>month</span>: 12</li></ul></div><ul class='xr-sections'><li class='xr-section-item'><div class='xr-array-wrap'><input id='section-e80aa26f-2658-4bf1-a674-d03f168a6dbc' class='xr-array-in' type='checkbox' checked><label for='section-e80aa26f-2658-4bf1-a674-d03f168a6dbc' title='Show/hide data repr'><svg class='icon xr-icon-database'><use xlink:href='#icon-database'></use></svg></label><div class='xr-array-preview xr-preview'><span>11 2 11 13 14 12 14 8 12 11 12 12 6 ... 9 13 13 5 14 12 14 9 12 9 17 8</span></div><div class='xr-array-data'><pre>array([[[11,  2, 11, ..., 11, 12, 12],
             [ 6, 10, 10, ..., 13, 14,  8],
             [ 9,  7,  9, ..., 10,  9, 10],
             ...,
@@ -537,9 +537,9 @@ Storing this dataset as an xarray data array
             ...,
             [12,  7,  5, ...,  9, 15,  5],
             [ 9,  5, 13, ..., 10, 11, 13],
[…]
