+++ stable/_sources/auto_examples/plot_split_half_analysis.rst.txt
@@ -330,7 +330,7 @@ further with five components. This indicates that three components is a good cho
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 1 minutes  3.490 seconds)
+   **Total running time of the script:** ( 1 minutes  11.130 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_split_half_analysis.py:
