+++ stable/_sources/auto_examples/plot_core_consistency.rst.txt
@@ -292,7 +292,7 @@ split half stability, and looking at the component vectors to see if they make s
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  10.288 seconds)
+   **Total running time of the script:** ( 0 minutes  11.676 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_core_consistency.py:
