+++ stable/_sources/auto_examples/plot_optimisation_diagnostic.rst.txt
@@ -452,7 +452,7 @@ Finally, let's compare our selected model with the true decomposition!
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  17.987 seconds)
+   **Total running time of the script:** ( 0 minutes  18.909 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_optimisation_diagnostic.py:
