+++ stable/_sources/auto_examples/plot_outlier_detection.rst.txt
@@ -473,7 +473,7 @@ Here, we see that we finally uncovered the true components. This is also apparen
 
  .. code-block:: none
 
-    The FMS is 0.92
+    The FMS is 0.95
 
 
 
@@ -629,7 +629,7 @@ to remove and not on a case-by-case basis.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  6.499 seconds)
+   **Total running time of the script:** ( 0 minutes  7.394 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_outlier_detection.py:
