### 2022-06-13 #4434 @faroit
👋 Hi @MarieRoald, thanks again for your submission to JOSS. I will be serving as the editor for this now. We'll use this issue to identify reviewers and resolve technical issues.

i won't be able to start the review process before start of next week as I am currently on vacation but I will already contact some scholars that I know have worked closely with parafac and python. @MarieRoald, please feel free to also suggest a couple of reviewers yourself.

For this, please check the JOSS [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) and only mention them with a space between the @ and their handle (so you would refer to me as "@ faroit").
