### 2022-06-21 #4434 @MarieRoald
Hi @faroit, neither @yngvem nor I have worked with @ aarmey directly, but all three of us are involved with the [TensorLy project](http://tensorly.org/stable/about.html). From the [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy), I am not sure if this is a COI or not?
