### 2022-06-21 #4434 @faroit
@MarieRoald to me this looks more like a COI than not... @arfon what do you think? The submission is part of the parent (tensorly) project and the potential reviewer (aarmey) is part of the parent project but hasn't contributed to this submission or has a direct working relationship with the submitting authors.
