### 2022-06-27 #4434 @MarieRoald
@faroit Update: We have now started discussing the possibility of more collaboration within the TensorLy team. While nothing is set in stone yet, I am now leaning more towards this being a COI.
