### 2022-07-02 56616ab
Clipping annotations within axes
