### 2022-08-09 3963411
Add HOTTBOX cite and minor changes
