+++ paper/paper.md
@@ -30,13 +30,13 @@ Multi-way data, also known as tensor data or data cubes, occur in many applicati
 
 # Statement of need
 
-Python has become a language of choice for data science for both industrial and academic research. Open source tools, such as scikit-learn [@pedregosa2011scikit] and Pandas [@mckinney-proc-scipy-2010] have made a variety of machine learning methods accessible within Python. Recently, TensorLy has also made tensor methods available in Python [@kossaifi2016tensorly], providing seamless integration of multi-way data mining methods within the python scientific environment. However, while TensorLy is an open-source, community-driven package for calculating tensor decompositions, it does not include tools for analysing or visualising the tensor decomposition models. Because tensor decompositions provide powerful tools to extract insight from multi-way data, effective visualisations are crucial as they are needed to communicate this insight. Furthermore, visualisation and evaluation are essential steps in the multi-way analysis pipeline — without tools for these steps, we cannot find suitable models.
+Python has become a language of choice for data science for both industrial and academic research. Open source tools, such as scikit-learn [@pedregosa2011scikit] and Pandas [@mckinney-proc-scipy-2010] have made a variety of machine learning methods accessible within Python. Recently, TensorLy has made tensor methods also available in Python [@kossaifi2016tensorly], providing seamless integration of multi-way data mining methods within the python scientific environment. However, while TensorLy is an open-source, community-driven package for calculating tensor decompositions, it does not include tools for analysing or visualising the tensor decomposition models. Because tensor decompositions provide powerful tools to extract insight from multi-way data, effective visualisations are crucial as they are needed to communicate this insight. Furthermore, visualisation and evaluation are essential steps in the multi-way analysis pipeline — without tools for these steps, we cannot find suitable models.
 
-There is, to our knowledge, no free open source software (FOSS) that facilitates all these steps. For MATLAB, some tools cover part of this scope, such as Tensor Toolbox (which provides some model evaluation) [@osti_1230898] or the N-Way toolbox [@andersson2000n]. PLSToolbox covers most of our scope, but it is a closed source commercial software. There is, therefore, a growing need for FOSS tools for the visualisation and evaluation of tensor decompositions.
+There is, to our knowledge, no free open source software (FOSS) that facilitates all these steps. Some tools cover part of this scope, such as Tensor Toolbox for MATLAB [@osti_1230898], which provides some functionality for model evaluation; and the N-Way toolbox for MATLAB[@andersson2000n] and Higher Order Tensor ToolBOX (HOTTBOX) for Python [@kisil2021hottbox], which both provide limited functionality for model evaluation and visualisation. Finally, PLSToolbox covers most of our scope, but it is a closed-source commercial software. There is, therefore, a growing need for FOSS tools for the visualisation and evaluation of tensor decompositions.
 
 # Example
 
-The PARAFAC model is straightforward, but there are several pitfalls to consider. Two main pitfalls are the scaling and permutation indeterminacy [@bro1997parafac]. The order of the components does not matter, and the magnitude of one factor matrix can be scaled arbitrarily so long as another factor matrix is inversely scaled (an even number of components may even change sign!). Therefore, it can be time-consuming and cumbersome to go from having fitted a PARAFAC model to visualising it. TLViz takes care of these hurdles in a transparent way. The code below shows how easy we can use TLViz and TensorLy to analyse a fluorescence spectroscopy dataset.
[…]
