+++ paper/paper.bib
@@ -121,3 +121,11 @@
   publisher = {Ubiquity Press},
   doi       = {10.5334/jors.148}
 }
+
+@article{kisil2021hottbox,
+  title     = {HOTTBOX: Higher Order Tensor ToolBOX},
+  author    = {Kisil, Ilya and Calvi, Giuseppe G and Dees, Bruno S and Mandic, Danilo P},
+  journal   = {arXiv preprint arXiv:2111.15662},
+  year      = {2021},
+  doi       = {10.48550/arXiv.2111.15662}
+}
