### 2022-08-11 #4434 @MarieRoald
@sara-02: Thank you, that's great!
@faroit: I believe we need at least two reviewers? I looked through the list again, and here are some additional candidates I found by broadening my search to people interested in data science/machine learning:

- canerturkmen
- 62442katieb
- jkbren
- TahiriNadia

And here are some more possible candidates (outside the list) familiar with Python and tensor decompositions:

- jvendrow
- IlyaKisil
- aschein
