### 2022-09-12 #4754 @editorialbot
[REVIEW]: TLViz: Visualising and analysing tensor decomposition models with Python
**Submitting author:** @marieroald (<a href="http://orcid.org/0000-0002-9571-8829">Marie Roald</a>)
**Repository:** https://github.com/tensorly/viz
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.6
