### 2022-09-12 #4754 @faroit
👏 @MarieRoald @@sara-02 @yiitozer - this is the review thread for the paper. All of our communications will happen here from now on.

```@editorialbot generate my checklist```

These checklists contain the JOSS requirements. As you go over the submission, please check any items that you feel have been satisfied. The first comment in this thread contains links to the JOSS reviewer guidelines.

The JOSS review is different from other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. The reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention [openjournals/joss-reviews/issues/4754](https://github.com/openjournals/joss-reviews/issues/4754) so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

__We aim for reviews to be completed within about 2-4 weeks__. Please let me know if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
