+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = tensorly-viz
-version = 0.1.4
+version = 0.1.5
 license = "MIT license",
 description = Package to visualise component-based decomposition models such as PCA and PARAFAC
 long_description = file: README.rst
