### 2022-10-24 tensorly/viz#7 @yiitozer
workflow runs failed
I could run all the tests successfully on my local machine following the instructions, but I have just seen that your last workflow run has failed. Is there a specific reason for this?

https://github.com/openjournals/joss-reviews/issues/4754
