### 2022-10-24 tensorly/viz#8 @yiitozer
error in the formula
In the documentation of [Core Consistency](http://tensorly.org/viz/stable/auto_examples/plot_core_consistency.html),

1) The following is written:
x_{ijk} is described by:
ijk = ...

whereas, I think on the LHS of the equation, it must be x_{ijk}.

2) Spelling error: compponent -> component

3) I would use a colorbar, where the interaction between different components are shown.

https://github.com/openjournals/joss-reviews/issues/4754
