### 2022-10-24 tensorly/viz#9 @yiitozer
ordering of the examples
In [TLViz + PlotLy for interactive visualisations](http://tensorly.org/viz/stable/auto_examples/plot_bike_plotly.html#tlviz-plotly-for-interactive-visualisations), the split-half analysis example is referred to, but it comes after the TLVir + PlotLy example. I would switch the ordering.

https://github.com/openjournals/joss-reviews/issues/4754
