### 2022-10-24 tensorly/viz#10 @yiitozer
plotting the factor similarity
In the figure of the section [Plotting the factor similarity](http://tensorly.org/viz/stable/auto_examples/plot_split_half_analysis.html), it would be nice to have axis labels. The x-axis is also a bit confusing since the number of components is supposed to be an integer, whereas we see decimal values here.

The same applies to the last figure. (Split-half analysis for the bike sharing data)

https://github.com/openjournals/joss-reviews/issues/4754
