### 2022-10-24 tensorly/viz#11 @yiitozer
plot optimisation diagnostic
The legends of the figures are cropped out in the documentation: [Optimisation diagnostics with PARAFAC models](http://tensorly.org/viz/stable/auto_examples/plot_optimisation_diagnostic.html)

https://github.com/openjournals/joss-reviews/issues/4754
