### 2022-10-24 tensorly/viz#12 @yiitozer
plot selecting aminoacids components
In this section, I could not understand the term *initialisation run*. Could you help me with this?

Please have a look at the figures in the "Core element plot"; some of the titles are not readable.

In the last paragraph, you might need to reformulate the following sentence:
"The data is samples from mixtures of three different aminoacids. "

https://github.com/openjournals/joss-reviews/issues/4754
