### 2022-10-25 5355381
Some enhancements of plotting functionality
