### 2022-10-25 c1dc1da
Update doctest to handle change in matplotlib string representation of axes
