### 2022-10-27 tensorly/viz#7 @yiitozer
Ah OK, I have overseen that. Thank you. I'm closing this issue.
