### 2022-10-27 bacfb25
Ignore visualisation doctests for python 3.7
