+++ .github/workflows/tests.yml
@@ -17,10 +17,14 @@ jobs:
         run: |
           python -m pip install --upgrade pip wheel
           python -m pip install -e .[test,all]
+      - name: run tests (python 3.7)
+        if: ${{matrix.python == '3.7'}}
+        run: python3 -m pytest --ignore tlviz/visualisation.py
       - name: run tests
-        run: python3 -m pytest --cov=tlviz --cov-report=xml
+        if: ${{matrix.python != '3.7'}}
+        run: python3 -m pytest --cov=tlviz
       - name: upload coverage to Codecov
-        if: ${{matrix.os == 'ubuntu-latest'}}
+        if: ${{matrix.os == 'ubuntu-latest' && matrix.python == '3.10'}}
         uses: codecov/codecov-action@v2
         with:
           verbose: true
