### 2022-10-27 d8c7494
Discontinue support for Python 3.7
