+++ .github/workflows/tests.yml
@@ -6,7 +6,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python: ['3.7', '3.8', '3.9', '3.10']
+        python: ['3.8', '3.9', '3.10']
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v2
@@ -17,11 +17,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip wheel
           python -m pip install -e .[test,all]
-      - name: run tests (python 3.7)
-        if: ${{matrix.python == '3.7'}}
-        run: python3 -m pytest --ignore tlviz/visualisation.py
       - name: run tests
-        if: ${{matrix.python != '3.7'}}
         run: python3 -m pytest --cov=tlviz
       - name: upload coverage to Codecov
         if: ${{matrix.os == 'ubuntu-latest' && matrix.python == '3.10'}}
