+++ dev/api/visualisation-3.py
@@ -4,5 +4,5 @@ import matplotlib.pyplot as plt
 cp_tensor = random_cp(shape=(5,10,15), rank=2)
 component_scatterplot(cp_tensor, mode=0)
 # Expected:
-## <AxesSubplot:title={'center':'Component plot'}, xlabel='Component 0', ylabel='Component 1'>
+## <AxesSubplot: title={'center': 'Component plot'}, xlabel='Component 0', ylabel='Component 1'>
 plt.show()
