### 2022-10-27 tensorly/viz#8 @MarieRoald
Closed by #13
