### 2022-10-27 tensorly/viz#12 @MarieRoald
Closed by #13
