### 2022-10-27 tensorly/viz#11 @MarieRoald
Closed by #13
