### 2022-10-27 tensorly/viz#10 @MarieRoald
Closed by #13
