### 2022-10-28 tensorly/viz#9 @yiitozer
I think that's a good solution and I also understand that it's not a step by step documentation. We can close this issue!
