### 2022-10-29 #4754 @sara-02
@faroit @MarieRoald
My minor comment on the paper
1. lines 49-54 in the Example section have inconsistent first-word capitalization in the lists. The end of lists can be converted to a full stop and then the start of the list can be capitalized.
