### 2022-10-30 #4754 @MarieRoald
@sara-02: Thank you so much for your comments! For the paper, I have updated the text to have capitalization and a full stop for each list element.
