### 2022-11-01 b511408
minor grammar improvements
