+++ paper/paper.md
@@ -36,7 +36,7 @@ There is, to our knowledge, no free open source software (FOSS) that facilitates
 
 # Example
 
-The PARAFAC model is straightforward, but there are several pitfalls to consider. Two main pitfalls are the scaling and permutation indeterminacy [@bro1997parafac]. The order of the components does not matter, and the magnitude of one factor matrix can be scaled arbitrarily so long as another factor matrix is inversely scaled (an even number of components may even change sign). Therefore, it can be time-consuming and cumbersome to go from having fitted a PARAFAC model to visualising it. TLViz takes care of these hurdles in a transparent way. The code below shows how easy we can use TLViz and TensorLy to analyse a fluorescence spectroscopy dataset.
+The PARAFAC model is straightforward, but there are several pitfalls to consider. Two main pitfalls are the scaling and permutation indeterminacy [@bro1997parafac]. The order of the components does not matter, and the magnitude of one factor matrix can be scaled arbitrarily so long as another factor matrix is inversely scaled (an even number of components may even change sign). Therefore, it can be time-consuming and cumbersome to go from having fitted a PARAFAC model to visualising it. TLViz takes care of these hurdles in a transparent way. The code below shows how easily we can use TLViz and TensorLy to analyse a fluorescence spectroscopy dataset.
 
 ```python
 import tlviz
@@ -91,13 +91,13 @@ TLViz follows the procedural paradigm, and all of TLViz’s functionality lies i
  1. `tlviz.utils` - general utilities that can be useful (e.g. forming dense tensors from decompositions)
  1. `tlviz.visualisation` - visualising component models
 
-A core design choice behind TLViz is how to store metadata. Consider the example above. It is necessary to know the values along the x-axis to interpret these components. Therefore, we use xarray DataArrays to store data tensors [@hoyer2017xarray], keeping the correct indices for each tensor mode (i.e. axis), and Pandas DataFrames to store factor matrices. However, TensorLy works with NumPy arrays. TLViz, therefore, provides useful tools to add the coordinates from an xarray DataArray onto the factor matrices obtained with TensorLy. Furthermore, all functions of TLViz support both labelled and unlabelled decompositions (i.e. DataFrames and NumPy arrays) and will use the labels whenever possible.
+A core design choice behind TLViz is how to store metadata. Consider the example above. It is necessary to know the values along the x-axis to interpret these components. Therefore, we use xarray DataArrays to store data tensors [@hoyer2017xarray], keeping the correct indices for each tensor mode (i.e. axis), and Pandas DataFrames to store factor matrices. However, TensorLy works with NumPy arrays. TLViz, therefore, provides useful tools to add the coordinates from an xarray DataArray onto the factor matrices obtained with TensorLy. Furthrmore, all functions of TLViz support both labelled and unlabelled decompositions (i.e. DataFrames and NumPy arrays) and will use the labels whenever possible.
 
 The visualisation module uses matplotlib to create the plots, and the goal of this module is to facilitate fast prototyping and exploratory analysis. However, TLViz can also seamlessly convert factor matrices into tidy tables, which are better suited for visualisation libraries such as Seaborn [@Waskom2021] and PlotLy Express, thus making it painless to combine tensor decomposition with the plotting library that best suits the user’s specific needs.
 
 To be easy to use, scientific software should have thorough and accurate documentation. For TLViz, this means two things: Explaining what the code does and why this is important. Therefore, we have taken care to review the literature, citing original sources wherever possible. By gathering the references together with the API documentation and examples, we make it straightforward for researchers new to the field to discover suitable references for their analysis.
 
[…]
