+++ stable/api/visualisation-9.py
@@ -6,5 +6,5 @@ true_cp, X = simulated_random_cp_tensor((10, 20, 30), 3, seed=0)
 est_cp = parafac(X, 3)
 histogram_of_residuals(est_cp, X)
 # Expected:
-## <AxesSubplot:title={'center':'Histogram of residuals'}, xlabel='Standardised residuals', ylabel='Frequency'>
+## <AxesSubplot: title={'center': 'Histogram of residuals'}, xlabel='Standardised residuals', ylabel='Frequency'>
 plt.show()
