### 2022-11-04 #4754 @MarieRoald
Great! Thanks @faroit! We have created a [new release (v0.1.6)](https://github.com/tensorly/viz/releases/tag/v0.1.6) and DOI: [10.5281/zenodo.7274925](https://doi.org/10.5281/zenodo.7274925)
